# Bug fix in the field map and G4 simulation performance improvement