--- a/packages/PHField/PHFieldSeaQuest.h
+++ b/packages/PHField/PHFieldSeaQuest.h
@@ -2,7 +2,7 @@
 #define __PHFieldSeaQuest_H__
 
 #include "PHField.h"
-#include "PHField3DCartesian.h"
+#include "SQField3DCartesian.h"
 #include "PHFieldRegionalConst.h"
 
 #include <map>
@@ -32,8 +32,8 @@ class PHFieldSeaQuest : public PHField
   float zValues[4];
   float kmagZOffset;
 
-  PHField3DCartesian fmag;
-  PHField3DCartesian kmag;
+  SQField3DCartesian fmag;
+  SQField3DCartesian kmag;
   PHFieldRegionalConst targetmag;
 };