… simulation/g4main/PHG4PhenixDetector.cc
… simulation/g4main/PHG4PhenixDetector.h
… simulation/g4main/PHG4PhenixSteppingAction.h
… simulation/g4main/PHG4Reco.h
… simulation/g4main/PHG4Subsystem.h