This PR consists of 2 parts: the first one is to fix a bug in the `PH3DFieldCartisian` class which was used for the field map interpolation in our core software; the second one is to improve the speed of G4 simulation via several changes. The original purpose of the PR was only to improve the simulation speed, however during the process I noticed the bug in the field map which caused the momentum reconstruction to be off by ~1%.

## Bug fix in the `PH3DFieldCartisian` 

In the current core software, the KMag and FMag are implemented by `packages/PHField/PHField3DCartesian`. This class read the text-based input grid file (contains x/y/z/Bx/By/Bz), save the content in a map using (x, y, z) as key and (Bx, By, Bz) as value. To obtain the magnetic field strength at a spacepoint (x, y, z), this class uses the map to locate the 8 grid points that are closest to the give spacepoint and performs a trilinear interpolation. There are two problems with this approach:
 1. using a map to locate the closest grid points, although being a o(log N) operation, is still very inefficient given that we are using fixed step size on x/y/z thus the index of the 8 closest grid points should be directly calculated using constant time;
 2. the implementation of the trilinear interpolation in `PHField3DCartesian` has a bug which have the upper-bound and lower-bound inverted. The difference is almost negligible inside the dipole magnets, but is noticeable in the fringe field region. The overall effect of this bug is that the field integral is increased by about 1%, which makes the reconstructed momentum smaller than the true momentum, as reported by Abi.

To fix this bug, I implemented another field map class `SQ3DFieldCartesian` under `package/PHField` with a hand-crafted trilinear interpolation algorithm (https://en.wikipedia.org/wiki/Trilinear_interpolation). The class `PHFieldSeaQuest` is also changed to use the new field map class.

## G4 simulation performance improvement

In addition to the more efficient field map interpolation method implemented above, this PR introduced 4 other small changes to speed up the simulation process.

  1. currently in addition to the default `"FTFP_BERT"` physics list, in the `PHG4Reco::InitRun()` optical photon process is also enabled by default, which is not useful in our simulation. I disabled this part by default and added a method `PHG4Reco::set_optical_photon(bool)` so that user can manually turn it back on;
  2. currently the simulation will track all particles till it's out-of-boundary or killed by Geant, a large amount of soft particles takes a lot of time to process. I added a function `PHG4Reco::set_energy_threshold(double)` to set the energy threshold for the G4 tracking, all the particles below this energy threshold will be killed without further propagation. This feature is turned off by default and user needs to explicitly set a threshold value to turn it on;
  3. currently we are using a global magnetic field manager even though the field map itself only covers down to z ~1550 cm. For tracks with z > 1550 cm the global field manager will always return 0 magnetic field but the RK4 propagation is still carried out for each step which generates a lot of useless computation. I created a zero-field-sub-world volume that contains all the downstream detectors with no magnetic field (all station-3 and station-4). The RK4 propagation is turned off in this volume to save time. This feature is turned off by default, user needs to call `PHG4Reco::set_zero_field_line(double)` to specify the beginning z position of the zero-field-sub-world to turn it on;
 4. currently at the beginning of the simulation, we read the text-based field map for FMag and KMag to create the look-up table for the field map interpolation. The text-based field map files are 59MB for FMag and 7MB for KMag. They take about 5 seconds to be loaded while running locally, which is not too bad. But when running tens of thousands of jobs on the grid, both field maps need to be uploaded to each node and thus constitute a significant overhead. In this PR I converted them to `TTree` and saved in a ROOT file, the size of the files are significantly reduced to 13MB and 1MB respectively. The local loading time dropped to about 1s. The old text-based field map is still supported but we should remove them from the resource package at a later stage.

## Tests

For all the tests below I generated 1500 single mu+ from `z = -300cm` with `-6GeV < px < 6GeV, -3GeV < py < 3GeV, 25GeV < pz < 100GeV`. The macro can be found here: https://github.com/liuk/e1039-analysis/blob/master/RecoDev/RecoE1039Sim.C.

### Field map bug fix

By changing to the new field map implementation, the average running time of `PHG4Reco` per event is reduced from **198ms** to **161ms**.  Below are the resolution of the momentum reconstruction at station-1 before/after the change. Left is before the change, right is after.

<img src="https://user-images.githubusercontent.com/4968504/86506598-6a09e980-bd96-11ea-8950-0e10c8917f06.png" alt="drawing" width="350"/><img src="https://user-images.githubusercontent.com/4968504/86506602-6e360700-bd96-11ea-8b06-42628021af7d.png" alt="drawing" width="350"/>

### Simulation speed improvement

On top of the test above, following three lines are added after line 124 of https://github.com/liuk/e1039-analysis/blob/master/RecoDev/RecoE1039Sim.C to enable the new speed improvement features in this PR:

```
g4Reco->set_optical_photon(false);   // disable the optical photon process
g4Reco->set_energy_threshold(0.0005);    // set the minimum energy threshold of particles to 0.0005GeV/0.5MeV
g4Reco->set_zero_field_line(1600);    // set the starting z position of the zero-field-sub-world to be 1600cm
```

The average time per event consumed by `PHG4Reco` reduced from **161ms** to **67ms**, the most significant contribution comes from the energy threshold cut. The momentum resolution plot below shows there is essentially no difference with these additional time-saving features.

<img src="https://user-images.githubusercontent.com/4968504/86507125-3a111500-bd9b-11ea-9b80-2e0f5c330cd1.png" width="350"/>