@abinashpun 2020-07-04
Dear Kun,

Thank you very much for the bug fix and improvement. I am able to build it successfully but I ran into a couple of problems while running the macro you mentioned. 
-I think one of them is due to inclusion of your anamodule ( I am working on building one).
-2nd is I got the following error about not finding Fmag.root.
----------------------------------------
_Error in <TFile::TFile>: file /e906/app/software/osg/software/e1039/resource/geometry/magnetic_fields/Fmag.root does not exist
SQField3DCartesian: cannot find the input ROOT field map, abort /e906/app/software/osg/software/e1039/resource/geometry/magnetic_fields/Fmag.root_
-------------------------------------