@abinashpun 2020-07-04
> > Dear Kun,
> > Thank you very much for the bug fix and improvement. I am able to build it successfully but I ran into a couple of problems while running the macro you mentioned.
> > 
> > * I think one of them is due to inclusion of your anamodule ( I am working on building one).
> > * 2nd is I got the following error about not finding Fmag.root.
> > 
> > ```
> > Error in <TFile::TFile>: file /e906/app/software/osg/software/e1039/resource/geometry/magnetic_fields/Fmag.root does not exist
> > SQField3DCartesian: cannot find the input ROOT field map, abort /e906/app/software/osg/software/e1039/resource/geometry/magnetic_fields/Fmag.root_
> > ```
> 
> Hello Abi, you do not really need to run my analysis module, it's only intended for myself to make quick QA plots. The updated core software should work with the old Fun4Sim macro with your own analysis module.
> 
> As for the field map, they are not updated on the official resource area yet, you can keep using the old tab.Fmag and tab.Kmag since they are sill supported in the new field map class. If you are interested in giving the ROOT field map a try, they can be found here: https://github.com/E1039-Collaboration/e1039-resource/tree/master/resource/geometry/magnetic_fields

Thank you, Kun. I am able to run the Fun4Sim successfully.