@knakano0524 2020-07-08
Sorry for my late response.  Both the two parts are certainly helpful.  Since Abi has already run the test, I just looked over the files changed, which seem fine.

I just wonder if we should set the energy threshold to non-zero **by default**.  Actually I had thought that it was already non-zero.  If we set it in macro, we will need an extra care to consistently update all macros.  As it is not urgent nor conclusive, you can decide it before merging this PR.