@liuk 2020-07-08
> Sorry for my late response. Both the two parts are certainly helpful. Since Abi has already run the test, I just looked over the files changed, which seem fine.
> 
> I just wonder if we should set the energy threshold to non-zero **by default**. Actually I had thought that it was already non-zero. If we set it in macro, we will need an extra care to consistently update all macros. As it is not urgent nor conclusive, you can decide it before merging this PR.

Hello Kenichi, thanks for the review and comment. I think it makes very good sense especially that in the default E906 simulation (GMC) a similar cut is applied at 1MeV. So this cut has been extensively tested in the muon simulation in some sense. I quickly run the test again using 1MeV as the threshold value and the result (running time and momentum resolution) is essentially the same. I will push a commit to turn this feature on in `PHG4Reco` with threshold set at 0.5MeV  by default.