@liuk 2020-07-09
Hello all, thank you both for the review. I will merge this change later today. Since this change contains a bug fix that impacts everyone, we should update the official repository ASAP. There should not be any impact to the user other than that the simulation is faster.