@abinashpun 2020-07-09
Dear Kun, 
Ok, I will send out the email and proceed to update our software version as soon as you merge it. 
Thank you.